97% done
Have to change Wallet when sold. To do that have to connect the sold button and the team self.wallet. Check out displaying the wallets separately in a different fn or class and get wallet data by getting the sum of players bought and subtracting. Or try locally changing the wallet.